Add files via upload
Requirements
Choose a React application for deployment. It can be a custom "hello world" application, or you can choose an open-source app from GitHub. If you choose the latter, ensure it is easy to work with (i.e., you can clone it locally and follow the instructions for building and running it locally). The React app should be self-contained, without any backend dependencies.
The application should be developed using TypeScript.
The app's features are not the focus of this assignment, but it should have at least one unit test.
Set up a CI/CD flow that includes transpiling, linting, and unit testing.
Choose either Azure DevOps or AWS CodePipeline for automating the CI/CD process.
Include code coverage reporting as part of the CI/CD flow.
Configure the process to fail if there are any issues with linting, transpiling, or unit testing.
The CI/CD flow should result in the React app being deployed to an Amazon S3 bucket. You can create the S3 bucket manually or use Terraform (bonus points). The S3 bucket must be configured for public access so the app can be accessed when deployed.
Demonstration